requirements for myamo radio.txt
Pillow
python-vlc
